+++ test-server/source/app/app-view.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import Component from '../../../build/component';
 import Link from '../../../build/link';
-import 'normalize.css/normalize.css!';
 
 export default class AppView extends Component {
   render() {